chore(docs): change npm install advice; solidify Meteor advice
It is my assessment that normal installs, and not complete installs, should be used for every day development. Yes, package lock updates are a side-effect, but they are a _desirable_ side-effect during development. We _do not_ want to be using outdated package resolution hashes. We _do not_ want to ignore inconsistences between the package and lock file (which "ci" errors on or ignores, but "i" actively fixes). 

We should establish _how to respond_ to package lock updates and encourage the universal practice.

Also, I removed some of the verbose or misleading language around Meteor. The situation is simpler and more logical (without outside citation needed) than initially present.